+++ docs/changelog.md
@@ -17,6 +17,7 @@ category: Administration
 * feat: improve log output for template import (#88)
 * docs: add documentation to enable informational logging (#88)
 * feat: refer to log files after template import (#88)
+* feat: provide workflow function to create subtasks (#34)
 
 ### [23.03.4] - 2023-03-29